+++ postgresstore/postgresstore.go
@@ -125,9 +125,11 @@ func (a *Store) GetInfo(ctx context.Context) (interface{}, error) {
 // NewBatch implements github.com/stratumn/go-core/store.Adapter.NewBatch.
 func (a *Store) NewBatch(ctx context.Context) (store.Batch, error) {
 	for b := range a.batches {
+		b.lock.RLock()
 		if b.done {
 			delete(a.batches, b)
 		}
+		b.lock.RUnlock()
 	}
 
 	tx, err := a.db.Begin()